Organize form components into own directories
