+++ src/components/forms/line/VehicleModeDropdown.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from 'react-i18next';
-import { ReusableComponentsVehicleModeEnum } from '../../generated/graphql';
-import { mapVehicleModeToUiName } from '../../i18n/uiNameMappings';
-import { FormInputProps } from '../../uiComponents';
-import { EnumDropdown } from './EnumDropdown';
+import { ReusableComponentsVehicleModeEnum } from '../../../generated/graphql';
+import { mapVehicleModeToUiName } from '../../../i18n/uiNameMappings';
+import { FormInputProps } from '../../../uiComponents';
+import { EnumDropdown } from '../common/EnumDropdown';
 
 interface Props extends FormInputProps {
   testId?: string;
