+++ src/components/forms/route/DirectionDropdown.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from 'react-i18next';
-import { mapDirectionToUiName } from '../../i18n/uiNameMappings';
-import { RouteDirection } from '../../types/RouteDirection';
-import { FormInputProps } from '../../uiComponents';
-import { EnumDropdown } from './EnumDropdown';
+import { mapDirectionToUiName } from '../../../i18n/uiNameMappings';
+import { RouteDirection } from '../../../types/RouteDirection';
+import { FormInputProps } from '../../../uiComponents';
+import { EnumDropdown } from '../common/EnumDropdown';
 
 interface Props extends FormInputProps {
   testId?: string;
