+++ src/hooks/routes/useEditRouteGeometry.ts
@@ -1,4 +1,4 @@
-import { RouteFormState } from '../../components/forms/RoutePropertiesForm.types';
+import { RouteFormState } from '../../components/forms/route/RoutePropertiesForm.types';
 import { RouteStop } from '../../context/MapEditor';
 import {
   InsertRouteOneMutationVariables,
