+++ src/hooks/useEditLine.ts
@@ -1,5 +1,5 @@
 import { useTranslation } from 'react-i18next';
-import { FormState } from '../components/forms/LineForm';
+import { FormState } from '../components/forms/line/LineForm';
 import {
   PatchLineMutationVariables,
   RouteLine,
